[CHORE] auth 폴더 구조 변경
