+++ src/controller/index.ts
@@ -1,3 +1,4 @@
 export { default as goalController } from "./goalController";
 export { default as homeController } from "./homeController";
 export { default as mypageController } from "./mypageController";
+export { default as authController } from "./authController";
